* main: (21 commits)
  suppress spotbugs warning RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE
  Add suppress warning for findbugs bug.
  Make mockable.
  Add missing pod for tests.
  Batch deletes with more than 1000 s3 Keys.
  Fix sql.
  Fix correctly this time.
  Woops result set adapter.
  Update sql.
  Actually use correct sql for deleting edges and fix binding.
  Add some additional logging.
  Ensure byteblob data manager is initialized.
  Bump hazelcast version.
  Move other interfaces to interface as they aren't getting picked up by post construct task.
  Woops. Weren't missing add logging.
  Add missing annotations.
  Delay hz map initialization until necessary.
  Inject required component.
  Don't require transient beans for initialization.
  Update column to be bigint (pg uses unsigned int)
  ...

+++ conductor-client/src/main/java/com/openlattice/postgres/PostgresColumn.java
@@ -387,7 +387,7 @@ public final class PostgresColumn {
     public static final String                   OID_FIELD                         = "oid";
     public static final PostgresColumnDefinition OID                               = new PostgresColumnDefinition(
             OID_FIELD,
-            INTEGER );
+            BIGINT );
     public static final String                   ORDINAL_POSITION_FIELD            = "ordinal_position";
     public static final PostgresColumnDefinition ORDINAL_POSITION                  =
             new PostgresColumnDefinition( ORDINAL_POSITION_FIELD, INTEGER ).notNull();
